refactor: split semesters.tsx into semesters.ts and sections.ts
